+++ packages/twenty-front/src/modules/settings/domains/components/SettingsWorkspaceDomainCard.tsx
@@ -1,4 +1,5 @@
 import { currentWorkspaceState } from '@/auth/states/currentWorkspaceState';
+import { isCloudflareIntegrationEnabledState } from '@/client-config/states/isCloudflareIntegrationEnabledState';
 import { isMultiWorkspaceEnabledState } from '@/client-config/states/isMultiWorkspaceEnabledState';
 import { SettingsCard } from '@/settings/components/SettingsCard';
 import { useLingui } from '@lingui/react/macro';
@@ -13,6 +14,9 @@ export const SettingsWorkspaceDomainCard = () => {
   const isMultiWorkspaceEnabled = useAtomStateValue(
     isMultiWorkspaceEnabledState,
   );
+  const isCloudflareIntegrationEnabled = useAtomStateValue(
+    isCloudflareIntegrationEnabledState,
+  );
   const currentWorkspace = useAtomStateValue(currentWorkspaceState);
 
   if (!isMultiWorkspaceEnabled) {
@@ -20,19 +24,26 @@ export const SettingsWorkspaceDomainCard = () => {
   }
 
   return (
-    <UndecoratedLink to={getSettingsPath(SettingsPath.Domain)}>
-      <SettingsCard
-        title={t`Customize Domain`}
-        Icon={<IconWorld />}
-        Status={
-          currentWorkspace?.customDomain &&
-          currentWorkspace?.isCustomDomainEnabled ? (
-            <Status text={t`Active`} color="turquoise" />
-          ) : currentWorkspace?.customDomain ? (
-            <Status text={t`Inactive`} color="orange" />
-          ) : undefined
-        }
-      />
-    </UndecoratedLink>
+    <>
+      <UndecoratedLink to={getSettingsPath(SettingsPath.Subdomain)}>
+        <SettingsCard title={t`Subdomain`} Icon={<IconWorld />} />
+      </UndecoratedLink>
+      {isCloudflareIntegrationEnabled && (
+        <UndecoratedLink to={getSettingsPath(SettingsPath.CustomDomain)}>
+          <SettingsCard
+            title={t`Custom Domain`}
+            Icon={<IconWorld />}
+            Status={
+              currentWorkspace?.customDomain &&
+              currentWorkspace?.isCustomDomainEnabled ? (
+                <Status text={t`Active`} color="turquoise" />
+              ) : currentWorkspace?.customDomain ? (
+                <Status text={t`Inactive`} color="orange" />
+              ) : undefined
+            }
+          />
+        </UndecoratedLink>
+      )}
+    </>
   );
 };
