+++ packages/twenty-shared/src/types/SettingsPath.ts
@@ -22,7 +22,8 @@ export enum SettingsPath {
   WorkspaceMemberPage = 'members/:workspaceMemberId',
   Workspace = 'general',
   Domains = 'domains',
-  Domain = 'domains/domain',
+  Subdomain = 'domains/subdomain',
+  CustomDomain = 'domains/custom-domain',
   PublicDomain = 'domains/public-domain',
   NewApprovedAccessDomain = 'domains/approved-access-domain/new',
   NewEmailingDomain = 'domains/emailing-domain/new',
